refactor: extract charge_fee to sablier-common crate
